docker/python/requirements.txt
docker==7.0.0
kafka-python==2.0.2
requests==2.31.0